Working DNS uprobes
