+++ contrib/amd64/bpf_tracing.h
@@ -1,0 +1,16 @@
+/* x86_64 (amd64) specific definitions for uprobes */
+#ifndef PT_REGS_DEFINED
+#define PT_REGS_DEFINED
+/* Architecture-specific register access macros */
+#define PT_REGS_PARM1(x) ((x)->di)
+#define PT_REGS_PARM2(x) ((x)->si)
+#define PT_REGS_PARM3(x) ((x)->dx)
+#define PT_REGS_PARM4(x) ((x)->cx)
+#define PT_REGS_PARM5(x) ((x)->r8)
+#define PT_REGS_PARM6(x) ((x)->r9)
+#define PT_REGS_RET(x) ((x)->sp)
+#define PT_REGS_FP(x) ((x)->bp)
+#define PT_REGS_RC(x) ((x)->ax)
+#define PT_REGS_SP(x) ((x)->sp)
+#define PT_REGS_IP(x) ((x)->ip)
+#endif /* PT_REGS_DEFINED */
