+++ contrib/arm64/bpf_tracing.h
@@ -1,0 +1,18 @@
+/* ARM64 specific definitions for uprobes */
+#ifndef PT_REGS_DEFINED
+#define PT_REGS_DEFINED
+/* Architecture-specific register access macros */
+#define PT_REGS_PARM1(x) ((x)->regs[0])
+#define PT_REGS_PARM2(x) ((x)->regs[1])
+#define PT_REGS_PARM3(x) ((x)->regs[2])
+#define PT_REGS_PARM4(x) ((x)->regs[3])
+#define PT_REGS_PARM5(x) ((x)->regs[4])
+#define PT_REGS_PARM6(x) ((x)->regs[5])
+#define PT_REGS_PARM7(x) ((x)->regs[6])
+#define PT_REGS_PARM8(x) ((x)->regs[7])
+#define PT_REGS_RET(x) ((x)->regs[30])
+#define PT_REGS_FP(x) ((x)->regs[29])
+#define PT_REGS_RC(x) ((x)->regs[0])
+#define PT_REGS_SP(x) ((x)->sp)
+#define PT_REGS_IP(x) ((x)->pc)
+#endif /* PT_REGS_DEFINED */
